Refactor element id dedupe logic into standalone module; dedupe element ids in pdf and epub
